Fix car search and rebuild carCriteriaBuilder and change package
